securisation suppr + validation comment + suppr article

+++ src/Controller/AdminController.php
@@ -9,8 +9,6 @@ use App\Repository\CommentRepository;
 use App\Repository\PostRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Bundle\SecurityBundle\Security;
-use Symfony\Component\Finder\Exception\AccessDeniedException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -19,8 +17,11 @@ use function Symfony\Component\Translation\t;
 
 class AdminController extends AbstractController
 {
+    private Request $request;
+
     public function __construct(private PostRepository $postRepository, private CommentRepository $commentRepository)
     {
+        $this->request = new Request();
     }
 
     #[Route('/admin', name: 'admin')]
@@ -42,7 +43,25 @@ class AdminController extends AbstractController
     public function postShow(): Response
     {
 
-        return $this->render('admin/post_show.html.twig', [
+        return $this->render('admin/post_admin.html.twig', [
+            'posts' => $this->postRepository->findAll()
+        ]);
+    }
+
+    #[Route('/admin/post/validate', name: 'admin_post_validate')]
+    #[IsGranted('ROLE_ADMIN', message: 'Accès refuser aux nom-admins', statusCode: 403)]
+    public function postValidate(Post $post): Response
+    {
+        $token = $this->request->query->get("token");
+
+        if ($this->isCsrfTokenValid('post' . $post->getId(), $token)) {
+
+            $post->setIsPublished(true);
+
+            $this->commentRepository->save($post, true);
+        }
+
+        return $this->render('admin/post_admin.html.twig', [
             'posts' => $this->postRepository->findAll()
         ]);
     }
@@ -52,17 +71,22 @@ class AdminController extends AbstractController
     public function postDelete(Post $post): Response
     {
 
-        $this->postRepository->remove($post, true);
+        $token = $this->request->query->get("token");
+
+        if ($this->isCsrfTokenValid('post' . $post->getId(), $token)) {
+            $this->postRepository->remove($post, true);
+        }
+
 
         return $this->redirectToRoute("admin_post_show");
     }
 
     #[Route('/admin/post/update/{id}', name: 'admin_post_update')]
     #[IsGranted('ROLE_ADMIN', message: 'Accès refuser aux nom-admins', statusCode: 403)]
-    public function postUpdate(Post $post, Request $request, EntityManagerInterface $entityManager): Response
+    public function postUpdate(Post $post, EntityManagerInterface $entityManager): Response
     {
         $updatePostForm = $this->createForm(PostFormType::class, $post);
-        $updatePostForm->handleRequest($request);
+        $updatePostForm->handleRequest($this->request);
 
 
         if ($updatePostForm->isSubmitted() && $updatePostForm->isValid()) {
@@ -79,36 +103,55 @@ class AdminController extends AbstractController
         ]);
     }
 
-    #[Route('/admin/comment/delete/{id}', name: 'admin_comment_delete')]
+    #[Route('/admin/comment/show/unvalid', name: 'admin_comment_unvalid')]
     #[IsGranted('ROLE_ADMIN', message: 'Accès refuser aux nom-admins', statusCode: 403)]
-    public function commentDelete(Post $post, Comment $comment): Response
+    public function showUnvalidComment(): Response
     {
+        $comments = $this->commentRepository->findBy(["isValid" => false]);
 
-        $this->commentRepository->remove($comment, true);
-
-        return $this->redirectToRoute('admin_comment_delete', ["id" => $post->getId()]);
-    }
-
-    #[Route('/admin/comment/show/unvalid/{id}', name: 'admin_comment_unvalid')]
-    #[IsGranted('ROLE_ADMIN', message: 'Accès refuser aux nom-admins', statusCode: 403)]
-    public function showUnvalidComment(Comment $comment): Response
-    {
-        ;
-
-        return $this->render('admin/comment_unvalid.html.twig', [
-            'comments' => $this->commentRepository->findBy(["isValid" => false])
+        return $this->render('admin/comment_admin.html.twig', [
+            'comments' => $comments
         ]);
     }
 
+    #[Route('/admin/comment/show/valid', name: 'admin_comment_valid')]
+    #[IsGranted('ROLE_ADMIN', message: 'Accès refuser aux nom-admins', statusCode: 403)]
+    public function showValidComment(): Response
+    {
+        $comments = $this->commentRepository->findBy(["isValid" => true]);
+
+        return $this->render('admin/comment_admin.html.twig', [
+            'comments' => $comments
+        ]);
+    }
+
+    #[Route('/admin/comment/delete/{id}', name: 'admin_comment_delete')]
+    #[IsGranted('ROLE_ADMIN', message: 'Accès refuser aux nom-admins', statusCode: 403)]
+    public function commentDelete(Comment $comment): Response
+    {
+        $token = $this->request->query->get("token");
+
+        if ($this->isCsrfTokenValid('comment' . $comment->getId(), $token)) {
+            $this->commentRepository->remove($comment, true);
+
+        }
+        return $this->redirectToRoute('admin');
+    }
 
     #[Route('/admin/comment/validate/{id}', name: 'admin_comment_validate')]
     #[IsGranted('ROLE_ADMIN', message: 'Accès refuser aux nom-admins', statusCode: 403)]
-    public function commentValidate(Post $post, Comment $comment): Response
+    public function commentValidate(Comment $comment): Response
     {
+        $token = $this->request->query->get("token");
 
-        $comment->setIsValid(true);
+        if ($this->isCsrfTokenValid('comment' . $comment->getId(), $token)) {
 
-        return $this->redirectToRoute('admin_comment_delete', ["id" => $post->getId()]);
+            $comment->setIsValid(true);
+
+            $this->commentRepository->save($comment, true);
+        }
+
+        return $this->redirectToRoute('admin_comment_unvalid');
     }
 
 }
